+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 - Script "XamlEdit" removed
 - Script "TestRunner" removed
+- Script "LoC" converted into an independent app
 
 ## 1.13.0 - 2017-02-17