+++ ChangeLog.md
@@ -1,10 +1,13 @@
-## 2.0.0 - 2020-08-
+## 2.0.0 - 2021-7-11
 
 - Script "XamlEdit" removed
 - Script "TestRunner" removed
 - Script "LoC" converted into an independent app
 - Loc XML output contains collected data recursively
-- Migrated to .Net Core
+- Migrated to .Net 5
+- WhiteRabbit:
+  - category removed
+  - monthly report added
 
 ## 1.13.0 - 2017-02-17